add glossary in card modal
