+++ json-altai/altai1.json
@@ -281,7 +281,7 @@
                 },
                 {
                     "id":"HAO-Q11-B",
-                    "title":"Did you take any specific oversight and control measures to reflect the =gb=interdependence=ge= of the system?",
+                    "title":"Did you take any specific oversight and control measures to reflect the interdependence of the system?",
                     "isVisibleIf":"",
                     "isScored":0,
                     "type":"radio",
