+++ package.json
@@ -49,6 +49,7 @@
     "react-mentions": "^4.4.7",
     "react-quill": "^1.3.5",
     "react-select": "^5.6.1",
+    "react-string-replace": "^1.1.1",
     "sass": "^1.62.1",
     "swr": "^1.3.0",
     "typed.js": "^2.0.15",
